[emacs compile] remove TERM env, don't apply ANSI colors
As of Emacs 27.1, the `compilation-filter-hook` is only called after the errors/warnings are parsed, so the ANSI
escape codes would be in the file path.

So, this prevents the compiler from outputting colors to the output!

In case the compiler doesn't honor TERM and still outputs the escape sequences, we can also advise
`compilation--ensure-parse` (or something) and remove the ANSI codes.